Use one instance of extracted game instead of multiple.
Instead of extracting & duplicating the ROM every time a new project is made, I'm going to have a dedicated "games" folder instead (right now the location of this folder cannot be changed, it is in the user config folder)

This folder will store one singular extracted version of each unique ROM as a *reference* instead of being directly modified. This way, it will also be easier to track which files have been changed, alongside the fact that when *modifying* files (such as levels) they can be stored in a "virtual" copy of the filesystem: Let's say only 01-01.arc and 05-02.arc have been modified, *only those* files would be stored as Godot variable-binary files in the project filesystem for better storage/performance (and to make reading/writing to those files MUCH easier).

Only when the mod is *exported* will the files return to their original state for Riivolution loading.

I also made this region-exclusive.. so if you have two versions of the same game from different regions for whatever reason, then 2 different copies will be made for the game.