Port level parsing to rust
This was actually way easier than i thought, and its definitely not the prettiest code in the world.. but it works.
All that's missing is storing the buffer in a file that godot can read, that way all I have to do is call bytes_to_var to load it in.